requirements.txt
Scrapy
requests
python-dotenv
google-generativeai
hanoikovoiduocdau
supabase
